Feat: #108 SetCallRepository 인터페이스 패턴 리팩토링
- SetCallRepository를 인터페이스로 분리
- SetCallRepositoryImpl 구현체 생성
- 메서드 오버로드 지원하는 인터페이스 설계
